fix(memory): purge relocated conversation-keyed tables in clearAll and prune-old-conversations (Codex P1)
